--- docker-compose.yml
@@ -1,8 +1,0 @@
-version: "3"
-
-services:
-  mailhog:
-    image: "mailhog/mailhog"
-    ports:
-      - "1025:1025"
-      - "8025:8025"
