+++ docs/getting-started/couriers.md
@@ -1,8 +1,8 @@
 Couriers are the drivers implementing the `Courier` interface. The following
 couriers are supported out of the box:
 
-* [Mail (using mail function)](/couriers/mail)
+* [Logging (test purposes only)](/couriers/logging)
+* [Mailjet](/couriers/mailjet)
 * [Postmark](/couriers/postmark)
 * [SendGrid](/couriers/sendgrid)
 * [SparkPost](/couriers/sparkpost)
-* [Mailjet](/couriers/mailjet)
